DEV: Change external_upload_stubs.filesize to bigint

+++ db/migrate/20210812033033_add_multipart_and_size_columns_to_external_upload_stubs.rb
@@ -4,7 +4,7 @@ class AddMultipartAndSizeColumnsToExternalUploadStubs < ActiveRecord::Migration[
   def up
     add_column :external_upload_stubs, :multipart, :boolean, default: false, null: false
     add_column :external_upload_stubs, :external_upload_identifier, :string, null: true
-    add_column :external_upload_stubs, :filesize, :integer
+    add_column :external_upload_stubs, :filesize, :bigint
 
     add_index :external_upload_stubs, :external_upload_identifier
